Make clang-tidy happy

+++ src/fheroes2/editor/editor_interface.cpp
@@ -1398,7 +1398,7 @@ namespace Interface
                         action.commit();
                     }
                     else if ( Artifact( static_cast<int>( objectInfo.metadata[0] ) ).isValid() ) {
-                        fheroes2::ArtifactDialogElement( objectInfo.metadata[0] ).showPopup( Dialog::OK );
+                        fheroes2::ArtifactDialogElement( static_cast<int>( objectInfo.metadata[0] ) ).showPopup( Dialog::OK );
                     }
                     else {
                         std::string msg = _( "%{object} has no properties to change." );
